+++ pkg/scheduler/src/lib.rs
@@ -12,6 +12,7 @@ pub mod preemption;
 pub mod scheduler;
 pub mod score;
 pub mod spread;
+pub mod virtualmachine;
 pub mod volumebinding;
 
 pub use scheduler::Scheduler;